Crate spl_associated_token_account

Convention for associating token accounts with a user wallet

Modules

error

Error types

instructionDeprecated

Program instructions

processor

Program state processor

tools

Utility functions

Constants

ID

The const program ID.

Functions

check_id

Returns true if given pubkey is the program ID.

create_associated_token_accountDeprecated

Create an associated token account for the given wallet address and token mint

get_associated_token_address

Derives the associated token account address for the given wallet address and token mint

get_associated_token_address_with_program_id

Derives the associated token account address for the given wallet address, token mint and token program id

id

Returns the program ID.