spl_associated_token_account

Function create_associated_token_account

pub fn create_associated_token_account(
    funding_address: &Pubkey,
    wallet_address: &Pubkey,
    token_mint_address: &Pubkey,
) -> Instruction

👎Deprecated since 1.0.5: please use instruction::create_associated_token_account instead

Create an associated token account for the given wallet address and token mint

Accounts expected by this instruction:

[writeable,signer] Funding account (must be a system account)

[writeable] Associated token account address to be created

[] Wallet address for the new associated token account

[] The token mint for the new associated token account

[] System program

[] SPL Token program