spl_associated_token_account::instruction

Enum AssociatedTokenAccountInstruction

pub enum AssociatedTokenAccountInstruction {
    Create,
    CreateIdempotent,
    RecoverNested,
}

Instructions supported by the AssociatedTokenAccount program

Variants

Create

Creates an associated token account for the given wallet address and token mint Returns an error if the account exists.

[writeable,signer] Funding account (must be a system account)

[writeable] Associated token account address to be created

[] Wallet address for the new associated token account

[] The token mint for the new associated token account

[] System program

[] SPL Token program

CreateIdempotent

Creates an associated token account for the given wallet address and token mint, if it doesn’t already exist. Returns an error if the account exists, but with a different owner.

[writeable,signer] Funding account (must be a system account)

[writeable] Associated token account address to be created

[] Wallet address for the new associated token account

[] The token mint for the new associated token account

[] System program

[] SPL Token program

RecoverNested

Transfers from and closes a nested associated token account: an associated token account owned by an associated token account.

The tokens are moved from the nested associated token account to the wallet’s associated token account, and the nested account lamports are moved to the wallet.

Note: Nested token accounts are an anti-pattern, and almost always created unintentionally, so this instruction should only be used to recover from errors.

[writeable] Nested associated token account, must be owned by 3

[] Token mint for the nested associated token account

[writeable] Wallet’s associated token account

[] Owner associated token account address, must be owned by 5

[] Token mint for the owner associated token account

[writeable, signer] Wallet address for the owner associated token account

[] SPL Token program