Crate spl_associated_token_account

Convention for associating token accounts with a user wallet

Re-exports

pub use solana_program;

Modules

instruction

Program instructions

processor

Program state processor

tools

Utility functions

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

create_associated_token_accountDeprecated

Create an associated token account for the given wallet address and token mint

get_associated_token_address

Derives the associated token account address for the given wallet address and token mint

id

Returns the program ID